WIP fix logging
