+++ alexa/package.json
@@ -13,27 +13,27 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@types/express": "^4.17.17",
-    "@types/i18next-sprintf-postprocessor": "^0.2.0",
-    "@types/lodash": "^4.14.198",
-    "@types/node": "^20.5.9",
-    "nodemon": "^3.0.1",
-    "rimraf": "^5.0.1",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.2.2"
+    "@types/express": "^5.0.3",
+    "@types/i18next-sprintf-postprocessor": "^0.2.3",
+    "@types/lodash": "^4.17.20",
+    "@types/node": "^24.0.10",
+    "nodemon": "^3.1.10",
+    "rimraf": "^6.0.1",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.8.3"
   },
   "dependencies": {
     "ask-sdk": "^2.14.0",
     "ask-sdk-core": "^2.14.0",
     "ask-sdk-express-adapter": "^2.14.0",
     "ask-sdk-model": "^1.86.0",
-    "axios": "^1.5.0",
-    "express": "^4.18.2",
-    "i18next": "^23.5.1",
+    "axios": "^1.10.0",
+    "express": "^5.1.0",
+    "i18next": "^25.3.1",
     "i18next-sprintf-postprocessor": "^0.2.2",
     "lodash": "^4.17.21",
-    "loglevel": "^1.8.1",
+    "loglevel": "^1.9.2",
     "loglevel-plugin-prefix": "^0.8.4",
-    "nano": "^10.1.2"
+    "nano": "^10.1.4"
   }
 }
