+++ client/package.json
@@ -31,11 +31,12 @@
     "jszip": "^3.10.1",
     "jwt-decode": "^4.0.0",
     "lodash-es": "4.17.21",
+    "loglevel": "^1.9.2",
     "loglevelnext": "6.0.0",
     "p-queue": "^8.1.0",
-    "pouchdb": "^8.0.1",
-    "pouchdb-core": "^8.0.1",
-    "pouchdb-selector-core": "^8.0.1",
+    "pouchdb": "^9.0.0",
+    "pouchdb-core": "^9.0.0",
+    "pouchdb-selector-core": "^9.0.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-i18next": "^15.6.0",
@@ -44,19 +45,7 @@
     "typescript": "^5.8.3",
     "use-pouchdb": "^2.0.2",
     "uuid": "^11.1.0",
-    "vite": "^7.0.2",
-    "workbox-background-sync": "^7.0.0",
-    "workbox-broadcast-update": "^7.0.0",
-    "workbox-cacheable-response": "^7.0.0",
-    "workbox-core": "^7.0.0",
-    "workbox-expiration": "^7.0.0",
-    "workbox-google-analytics": "^7.0.0",
-    "workbox-navigation-preload": "^7.0.0",
-    "workbox-precaching": "^7.0.0",
-    "workbox-range-requests": "^7.0.0",
-    "workbox-routing": "^7.0.0",
-    "workbox-strategies": "^7.0.0",
-    "workbox-streams": "^7.0.0"
+    "vite": "^7.0.2"
   },
   "scripts": {
     "ionic:serve:before": "bin/updateenv.sh && cp env-config.js ./public/env-config.js",
@@ -102,8 +91,8 @@
     "@types/react-router": "^5.1.11",
     "@types/react-router-dom": "^5.1.7",
     "@types/uuid": "^10.0.0",
-    "pouchdb-browser": "^8.0.1",
-    "pouchdb-find": "^8.0.1"
+    "pouchdb-browser": "^9.0.0",
+    "pouchdb-find": "^9.0.0"
   },
   "description": "An Ionic project"
 }
