+++ client/src/components/RemoteUtilities.tsx
@@ -11,7 +11,7 @@ import { urlPatternValidation, usernamePatternValidation, emailPatternValidation
 import { cloneDeep, pick, keys, isEqual } from 'lodash-es';
 import { Device } from '@capacitor/device';
 import { t } from "i18next";
-import { log } from "./Utilities";
+import log from "./logger";
 
 export async function getDeviceID() : Promise<string> {
     const devIDInfo = await Device.getId();
