+++ client/src/pages/ConflictLog.tsx
@@ -10,7 +10,7 @@ import ErrorPage from './ErrorPage';
 import { Loading } from '../components/Loading';
 import { ConflictDocs } from '../components/DBSchema';
 import { useTranslation } from 'react-i18next';
-import { log } from "../components/Utilities";
+import log from "../components/logger";
 
 const ConflictLog: React.FC<HistoryProps> = (props: HistoryProps) => {
   const { setDBCredsValue } = useContext(RemoteDBStateContext);
